Renamed `AlbumGroup` to `AlbumType`.
Added documentation about how omitting the market parameter when using the client credentials flow causes episodes and shows to not be returned.

Added `SpotifyAPI.filteredPlaylistItems(_:filters:additionalTypes:limit:offset:market:)`.

Renamed `SpotifyAPI.filteredPlaylistRequest(_:filters:additionalTypes:market:)` to `SpotifyAPI.filteredPlaylist(_:filters:additionalTypes:market:)`.

Renamed `SpotifyAPI.getPlaylistImage(_:)` to `SpotifyAPI.playlistImage(_:)`.

Added `snapshotId` parameter to `SpotifyAPI.removeAllOccurencesFromPlaylist(_:of:snapshotId:)`

Added `SpotifyPlayerError`; This error object is returned by Spotify when there are errors related to the player endpoints.

Fixed issues with additional types parameter for the playlist endpoints and added clearer documentation to `filteredPlaylist` and `filteredPlaylistItems`.

The generic `Item` type of `PlaylistItemContainer` is now optional because the episodes in a playlist will be `nil` If they were retrieved using the client credentials flow manager and a value for the `market` parameter was not provided.

Added a jpeg image to the `SpotifyExampleContent` module.

Renamed `AlbumGroup` to `AlbumType`.

Fixed bug with decoding of `Album` where both the `albumGroup` and `albumType` properties were being decoded from the `albumGroup` JSON key.

Fixed bug where the incorrect coding key was being used for the `trackNumber` and `previewURL` properties of `Track`, causing them to never be decoded from the data and always set to `nil`.

Fixed bug where `externalURLs` property of `SpotifyContext` was not being decoded because the JSON key name was incorrect.

Internal

Renamed `getPlaylistItems` to `_playlistItems`.

Renamed `URIsDict` to `URIsContainer` and made it public.

URIsContainer now contains an array of `SpotifyURIConvertible` instead of `[[String: String]]`.

`URIWithPositions.uri` is now `SpotifyURIConvertible` instead of `String`.

Ensured that passing in an empty array for `additionalType` to playlist methods does not cause the value for the query parameter to be set to an empty string. Instead, the query parameter is omitted from the URL.

`AudioAnalysis` now uses `Section` instead of `[[String: Double]]`, like it should've in the first place.

Removed `acceptApplicationJSON` and added `contentTypeJSON`. The former wasn't necessary as Spotify returns all data in JSON format. The Spotify web API documentation does, however, say that the "Content-Type: application/json" header should be used for all requests that contain JSON in the body, even though omitting this parameter doesn't result in an error. Therefore this header is now used as necessary. Also changed `bearerAuthorizationAndAcceptApplicationJSON` to `bearerAuthorizationAndContentTypeJSON`.

Added approximate equality methods for many types with floating point properties.

`anyFailingPublisher` can now infer the `Output` type of the publisher.

`SpotifyDecodingError` decoding error now saves the raw data AND the pretty error message to a folder when a decoding error occurs.

`SpotifyError` is now `Encodable`, whereas previously it was only `Decodable`.

Added Track tests, more playlist and player tests, and tests for playlist endpoints documentation.