Added documentation to search endpoint about how the market parameter is required for retrieving shows and episodes if the application has been authorized using the client credentials flow.
The type of `SearchResult.episodes` has been changed from `PagingObject<Episode>?` to `PagingObject<Episode?>?` and The type of `shows` has been changed from `PagingObject<Show>?` to `PagingObject<Show?>?`. This change was necessary because Spotify can return nil for these properties if the shows and/or episodes are not available in the specified markets.