Added documentation to the `shows(_:market)` and `episodes(_:market)` endpoints about how providing a single invalid id causes the entire request to fail, unlike the tracks, albums, and artists, endpoints.
Fixed bugs for `showEpisodes` endpoint which was trying to decode the wrong response type and did not add the id of the show to the query string correctly.

Added `totalEpisodes` property to `Show`, which should've already been added.

Added tests for the show endpoints.

Regenerated some of the example json objects and added the example playlistItems to the SpotifyExampleContent module.