Fixed a bug where the values passed into `SpotifyLocalError.invalidState` were transposed.
Simplified the documentation for the `authorizationManagerDidChange`.

Renamed `removeIfNil` to `urlQueryDictionary`.

Updated documentation for playing content to mention that shows can be used for `contextURI`.

Each method that causes a signal to be sent to `didChange` has documentation that mentions it.

Simplified the `localizedDescription` for the error objects. The descriptions are now more suitable for the end user. Always use the string representation of the error objects for debugging purposes; only use `localizedDescription` to display the error to the end user.

Added tests for the episodes and user profile methods.