Removed the `showDialog` parameter from `AuthorizationCodeFlowPKCEManager.makeAuthorizationURL(redirectURI:codeChallenge:state:scopes:)` because it is not actually supported by the Authorization Code Flow with Proof Key for Code Exchange. It is only supported by the Authorization Code Flow.
Fixed a bug in which the port, username, and password components of a URL were removed by the `URL.removingQueryItems()` method, when it should've only removed the query items and fragment. This caused the methods for requesting access and refresh tokens to fail when the redirect URI had these components. The initializers for `URL` and `URLComponents` now accept a port.

Removed the period and tilde characters from `String.urlSafeCharacters` because they are reserved in certain components of a URL (although not in the query string, where they will most likely be used).

Added server for listening for redirects after authorization.

Changed redirect URI for tests from http://localhost to http://localhost:8080.

Replaced `HTTTURLResponse.value(forHTTPHeaderField:)` with `HTTTURLResponse.allHeaderFields` because it is not available in certain versions of linux.

Fixed a bug in which `expectation.fulfill` in the tests was not being called on all code paths in multiple cases. This lead to the tests freezing until the expectations timed out.

publicly exposed `setExpirationDate` for the authorization managers.