Added Method for disabling the bootstrapping of `SpotifyAPILogHandler`: `SpotifyAPILogHandler.disable`.
`SpotifyAPI.search` now throws an error if `categories` is empty.

When comparing the authorization managers for equality, an absolute tolerance of 1 second (instead of three seconds) is used when comparing the expiration dates.

Added a new protocol, `ApproximatelyEquatable`, which is conformed too by all types in the object model that have Date or floating point properties or that contain other types with those properties.

Made `CurrentlyPlayingContext.logger` public; it should have always been public.

Fixed a bug in `AuthorizationCodeFlowManager.==` and `AuthorizationCodeFlowPKCEManager.==`: The access tokens were being compared twice and the refresh tokens weren't being compared.

Refactored the equatable and hashable methods of `AuthorizationCodeFlowManager` and `AuthorizationCodeFlowPKCEManager`.

Refactored the Codable methods for `PlaybackRequest` and `OffsetOption`.

Added tests for decoding `SpotifyError` and `SpotifyPlayerError`.

Regenerated Docs and linuxmain.