Added `SpotifyLocalError.httpError(HTTPURLResponse, Data)`. This error is returned when the status code of the response from the server is in the 4xx or 5xx range and the response body could not be decoded into any of the other errors types (`SpotifyAuthenticationError`, `SpotifyError`, `SpotifyPlayerError`).
A request will be automatically retried up to three times if it returns a `SpotifyLocalError.httpError` with a status code of 500, 502, 503, or 504.

When decoding the data from a request into a Swift type, the data will *first* be decoded into an error object if the status code is in the 4xx or 5xx range.

Internal

Added `SpotifyExampleImages.annabelleTooLarge`.

The methods for refreshing the tokens subscribe on a serial queue, which can prevent certain issues associated with concurrent queues.