Changed tests to be more generic over authorization manger backends and to prepare for testing with multiple backends.
Added loggers for backends.

The requirements for the backend protocols which make the token requests are now throwing.

Made `SpotifyAuthenticationError.errorDescription` optional. It can be nil if a request is sent with empty strings for the client id and client secret.

Made some of the objects in "Sources/SpotifyWebAPI/Object Model/Authentication Objects" public.

Fixed bug in `Dictionary.formURLEncoded` which used the non-percent-encoded query string to create the form-url-encoded data, which would caused problems with percent encoded strings.

Removed dangling function in `SpotifyUser.init`