First commit towards 2.0 release.
Removed ALL of the deprecated symbols.

Publisher extensions where the output is `URLResponse` are now extensions where the output is `HTTPURLResponse`. This removes the need for `castToURLResponse`. Other methods that returned `URLResponse`, such as the wrappers for `SpotifyAPI`, now also return `HTTPURLResponse`.

Added `type` property to `Playlist`.

Refactored `ContextOption` and `OffsetOption` as nested types under `PlaybackRequest`.